requirements.txt
pyhtcc
paho-mqtt